Adding explicit cast to FDateTime::FromUnixTimestampDecimal
[CL 26171989 by william cork in 5.3 branch]

+++ Engine/Source/Runtime/Core/Public/Misc/DateTime.h
@@ -631,7 +631,7 @@ public:
 	 */
 	static FDateTime FromUnixTimestampDecimal(double UnixTime)
 	{
-		return FDateTime(1970, 1, 1) + FTimespan(UnixTime * ETimespan::TicksPerSecond);
+		return FDateTime(1970, 1, 1) + FTimespan(int64(UnixTime * ETimespan::TicksPerSecond));
 	}
 
 	/**
